+++ src/i18n/zh/events.js
@@ -2,4 +2,6 @@ import { eventsA } from './events-a.js';
 import { eventsB } from './events-b.js';
 import { eventsC } from './events-c.js';
 import { eventsD } from './events-d.js';
-export const events = { ...eventsA, ...eventsB, ...eventsC, ...eventsD };
+import { eventsE } from './events-e.js';
+import { eventsF } from './events-f.js';
+export const events = { ...eventsA, ...eventsB, ...eventsC, ...eventsD, ...eventsE, ...eventsF };